requirements.txt
beautifulsoup4
pandas
requests
scikit-learn
matplotlib
seaborn
causalml
streamlit
uvicorn
fastapi
pydantic
mlflow
shap
numpy
python-dotenv
huggingface_hub